[rename/#6] - 채팅 관련 아이콘 채팅 폴더로 이동 및 import 변경
